+++ modules/crawler/tests/data/configs/crawler.php
@@ -1,8 +1,0 @@
-<?php
-
-return [
-    'id' => 'testenv',
-    'siteTitle' => 'Luya Tests',
-    'remoteToken' => 'testtoken',
-    'basePath' => dirname(__DIR__),
-];